pyproject.toml
[tool.poetry]
name = "iterm-simulator"
version = "0.1.0"
description = ""
authors = ["Mark Needham <m.h.needham@gmail.com>"]
readme = "README.md"
packages = [{include = "iterm_simulator"}]

[tool.poetry.dependencies]
python = "^3.11"
pip = "^23.2.1"
install = "^1.3.5"
iterm2 = "^2.7"
markdown-it-py = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
